+++ resources/views/livewire/admin/designation-component.blade.php
@@ -60,13 +60,16 @@
                                         @endif
                                         <td>{{ date('d-m-Y', strtotime($designation->created_at)) }}</td>
                                         <td class="table-action">
-                                            <a href="javascript: void(0);" class="action-ico btn btn-outline-info mx-1"> <i
-                                                    class="bi bi-pencil-square" data-bs-toggle="modal"
+                                            <a href="javascript: void(0);" class="action-ico btn btn-outline-info mx-1">
+                                                <i class="bi bi-pencil-square" data-bs-toggle="modal"
                                                     wire:click="editdata({{ $designation->id }})"
                                                     data-bs-target="#editdesignation"></i></a>
-                                            <a href="javascript: void(0);"
-                                                wire:click="deleteConfirmation({{ $designation->id }})"
-                                                class="action-ico btn btn-outline-danger mx-1"> <i class="bi bi-trash"></i></a>
+                                            @if (Auth::user()->hasPermission(['master-access']))
+                                                <a href="javascript: void(0);"
+                                                    wire:click="deleteConfirmation({{ $designation->id }})"
+                                                    class="action-ico btn btn-outline-danger mx-1"> <i
+                                                        class="bi bi-trash"></i></a>
+                                            @endif
                                         </td>
                                     </tr>
                                 @endforeach
